+++ src/main/java/kr/co/hospital/HospitalApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package kr.co.hospital;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
