+++ src/main/java/kr/co/hospital/SitemeshConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package kr.co.hospital;
 
 import org.sitemesh.builder.SiteMeshFilterBuilder;
 import org.sitemesh.config.ConfigurableSiteMeshFilter;
