+++ src/main/java/kr/co/hospital/admin/mapper/AdminMainMapper.java
@@ -1,11 +1,11 @@
-package com.example.demo.admin.mapper;
+package kr.co.hospital.admin.mapper;
 
 import java.util.ArrayList;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import com.example.demo.admin.dto.AdminTestDto;
-import com.example.demo.admin.dto.DoctorDto;
+import kr.co.hospital.admin.dto.AdminTestDto;
+import kr.co.hospital.admin.dto.DoctorDto;
 
 @Mapper
 public interface AdminMainMapper {
