+++ src/main/java/kr/co/hospital/client/controller/ClientMainController.java
@@ -1,11 +1,11 @@
-package com.example.demo.client.controller;
+package kr.co.hospital.client.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.example.demo.client.service.ClientMainService;
+import kr.co.hospital.client.service.ClientMainService;
 
 @Controller
 public class ClientMainController {
