+++ src/main/resources/mapper/client/ClientMainMapper.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.demo.client.mapper.ClientMainMapper">
-  <select id="clienttest" resultType="com.example.demo.client.dto.ClientTestDto">
-    select * from test
-  </select>
+<mapper namespace="co.kr.hospital.client.mapper.ClientMainMapper">
+  
 </mapper>
